### 2024-04-12 #168 @isabelizimm
Hello there! To give some clarity on what is the happening behind the scenes, pyOpenSci is currently updating its scope around analytic/modeling packages, which will affect the decision around QuadratiK. This is something that we want to do thoughtfully, to make sure we have clear guidelines and support on what that scope should be.

Thank you so much for your patience and your submission! We are wrapping up this process and hope to have our updated scope and decision on Quadratik to share with you soon.
