### 2024-04-29 61ca4a8
Updated _cv_functions.py, _h_selection.py, _kernel_test.py, kernel_test/_utils.py, _poisson_kernel_test.py
Incorporated changes in the test statistics (included U-statistics and V-statistics for the test), changed Uniformity Test Statistic, removed latex formatting from the h_selection plots, completed black formatting.
