+++ QuadratiK/kernel_test/_kernel_test.py
@@ -97,49 +97,49 @@ class KernelTest:
 
     Examples
     --------
-    >>> # Example for normality test
     >>> import numpy as np
+    >>> np.random.seed(78990)
     >>> from QuadratiK.kernel_test import KernelTest
-    >>> np.random.seed(42)
-    >>> data = np.random.randn(100,5)
-    >>> normality_test = KernelTest(h=0.4, centering_type="param",random_state=42).test(data)
-    >>> print("Test : {}".format(normality_test.test_type_))
-    >>> print("Execution time: {:.3f}".format(normality_test.execution_time))
-    >>> print("H0 is Rejected : {}".format(normality_test.h0_rejected_))
-    >>> print("Test Statistic : {}".format(normality_test.test_statistic_))
-    >>> print("Critical Value (CV) : {}".format(normality_test.cv_))
-    >>> print("CV Method : {}".format(normality_test.cv_method_))
-    >>> print("Selected tuning parameter : {}".format(normality_test.h))
+    >>> # data generation
+    >>> data_norm = np.random.multivariate_normal(mean = np.zeros(4), cov = np.eye(4),size = 500)
+    >>> # performing the normality test
+    >>> normality_test = KernelTest(h=0.4, num_iter=150, method= "subsampling", random_state=42).test(data_norm)
+    >>> print(f"Test : {normality_test.test_type_}")
+    >>> print(f"Execution time: {normality_test.execution_time:.3f}")
+    >>> print(f"H0 is Rejected : {normality_test.un_h0_rejected_}")
+    >>> print(f"Test Statistic : {normality_test.un_test_statistic_}")
+    >>> print(f"Critical Value (CV) : {normality_test.un_cv_}")
+    >>> print(f"CV Method : {normality_test.cv_method_}")
     ... Test : Kernel-based quadratic distance Normality test
-    ... Execution time: 0.096
+    ... Execution time: 0.356
     ... H0 is Rejected : False
-    ... Test Statistic : -8.588873037044384e-05
-    ... Critical Value (CV) : 0.0004464111809800183
+    ... Test Statistic : 0.01018599246239244
+    ... Critical Value (CV) : 0.07765034009837886
     ... CV Method : Empirical
-    ... Selected tuning parameter : 0.4
 
-    >>> # Example for two sample test
     >>> import numpy as np
+    >>> np.random.seed(0)
+    >>> from scipy.stats import skewnorm
     >>> from QuadratiK.kernel_test import KernelTest
-    >>> np.random.seed(42)
-    >>> X = np.random.randn(100,5)
-    >>> np.random.seed(42)
-    >>> Y = np.random.randn(100,5)
-    >>> two_sample_test = KernelTest(h=0.4, centering_type="param").test(X,Y)
+    >>> # data generation
+    >>> X_2 = np.random.multivariate_normal(mean = np.zeros(4), cov = np.eye(4), size=200)
+    >>> Y_2 = skewnorm.rvs(size=(200, 4),loc=np.zeros(4), scale=np.ones(4),a=np.repeat(0.5,4), random_state=0)
+    >>> # performing the two sample test
+    >>> two_sample_test = KernelTest(h = 2,num_iter = 150, random_state=42).test(X_2,Y_2)
     >>> print("Test : {}".format(two_sample_test.test_type_))
-    >>> print("Execution time: {:.3f}".format(two_sample_test.execution_time))
-    >>> print("H0 is Rejected : {}".format(two_sample_test.h0_rejected_))
-    >>> print("Test Statistic : {}".format(two_sample_test.test_statistic_))
-    >>> print("Critical Value (CV) : {}".format(two_sample_test.cv_))
[…]
