### 2024-04-29 4f0648a
Updated tests, UI, removed redundant import & latex from h_selection
