+++ QuadratiK/ui/pages/1_Normality_Test.py
@@ -76,12 +76,23 @@ if data is not None:
             res = pd.DataFrame()
             res["Value"] = [
                 norm_test.test_type_,
-                norm_test.test_statistic_,
-                norm_test.cv_,
-                norm_test.h0_rejected_,
+                norm_test.un_test_statistic_,
+                norm_test.un_cv_,
+                norm_test.un_h0_rejected_,
+                norm_test.vn_test_statistic_,
+                norm_test.vn_cv_,
+                norm_test.vn_h0_rejected_,
             ]
             res = res.set_axis(
-                ["Test Type", "Test Statistic", "Critical Value", "Reject H0"]
+                [
+                    "Test Type",
+                    "Un Test Statistic",
+                    "Un Critical Value",
+                    "Un Reject H0",
+                    "Vn Test Statistic",
+                    "Vn Critical Value",
+                    "Vn Reject H0",
+                ]
             )
             st.table(res)
             csv_res = res.to_csv().encode()
