### 2024-05-02 #168 @Batalex
Hello @rmj3197! I am Alex, and I have taken the Editor-in-Chief mantel for now!
We have decided that `QuadratiK` is in scope for us. I just want to make one thing clear, as the documentation highlights the novelty of the clustering algorithm, that our expertise is focused on good development and packaging practices, and not so much as a technical endorsement of the approach.
If that works for you, you are welcome to open a new issue referencing this pre-submission inquiry. Thank you for your patience.
