+++ QuadratiK/kernel_test/_utils.py
@@ -104,6 +104,141 @@ def param_centering(kmat_zz, z_mat, cov_h, mu_hat, sigma_hat):
     return k_center
 
 
+# def dof_normality_test(d, h):
+#     numerator = 1 - (h**2 / (h**2 + 2)) ** (d / 2)
+#     denominator = (
+#         (h**2 + 2) / (h**2 + 4) ** (d / 2)
+#         - 2 * (h**2 / (h**2 + 1)) ** (d / 2) * (h**2 / (h**2 + 3)) ** (d / 2)
+#         + (h**2 / (h**2 + 2)) ** d
+#     )
+#     dof = numerator**2 / denominator
+#     const = (
+#         ((h**2 + 2) / (h**2 * (h**2 + 4))) ** (d / 2)
+#         - 2 * (h**2 / (h**2 + 1)) ** (d / 2) * (1 / (h**2 + 3)) ** (d / 2)
+#         + (h / (h**2 + 2)) ** (d)
+#     ) / numerator
+
+#     return (dof, const)
+
+
+def dof_normality_test(sigma_h, v):
+    num_dof = np.linalg.det(sigma_h) ** (-1 / 2) - np.linalg.det(sigma_h + 2 * v) ** (
+        -1 / 2
+    )
+    den_dof = (
+        np.linalg.det(sigma_h) ** (-1 / 2) * np.linalg.det(sigma_h + 4 * v) ** (-1 / 2)
+        - 2
+        * np.linalg.det(sigma_h + v) ** (-1 / 2)
+        * np.linalg.det(sigma_h + 3 * v) ** (-1 / 2)
+        + np.linalg.det(sigma_h + 2 * v) ** (-1 / 2)
+    )
+
+    dof = num_dof**2 / den_dof
+    const = den_dof / num_dof
+
+    return (dof, const)
+
+
+# def variance_normality_test(d, h, n):
+#     var = (
+#         2
+#         / (n * (n - 1))
+#         * 1
+#         / (2 * np.pi) ** d
+#         * ((h**2 + 2) / (h**2**2 * (h**2 + 4))) ** (d / 2)
+#         - 2 / ((h**2 + 1) * (h**2 + 3)) ** (d / 2)
+#         + (h**2 + 2) ** (-d)
+#     )
+#     return var
+
+
+def variance_normality_test(sigma_h, v, n):
+    d = sigma_h.shape[0]
+    var = (
+        2
+        / (n * (n - 1))
+        * (1 / ((2 * np.pi) ** d))
+        * (
+            np.linalg.det(sigma_h) ** (-1 / 2)
+            * np.linalg.det(sigma_h + 4 * v) ** (-1 / 2)
+            - 2
+            * np.linalg.det(sigma_h + v) ** (-1 / 2)
+            * np.linalg.det(sigma_h + 3 * v) ** (-1 / 2)
+            + np.linalg.det(sigma_h + 2 * v) ** (-1)
+        )
+    )
+    return var
+
+
+def variance_two_sample_test(k_cen, n, m):
+    np.fill_diagonal(k_cen, 0)
+
+    K_xx = k_cen[:n, :n]
+    K_yy = k_cen[n : n + m, n : n + m]
+    K_xy = k_cen[:n, n : n + m]
+
+    # Factors
+    n_factor = 1.0 / (n * (n - 1))
+    m_factor = 1.0 / (m * (m - 1))
+    cross_factor = 1.0 / (n * m)
+
+    # Variance estimate calculation
+    est_var_D = (
+        2 * n_factor**2 * (K_xx**2).sum()
+        + 8 * cross_factor**2 * (K_xy**2).sum()
[…]
