+++ QuadratiK/kernel_test/_h_selection.py
@@ -110,7 +110,8 @@ def _objective_one_sample(
         random_state=random_state,
     )
 
-    statistic = stat_normality_test(xnew, h, np.array([mean_dat]), np.diag(s_dat))
+    statistic = stat_normality_test(
+        xnew, h, np.array([mean_dat]), np.diag(s_dat))
     cv = cv_normality(
         n,
         h,
@@ -232,7 +233,7 @@ def _objective_two_sample(
         skew_tilde = skew_data + dk
         s_tilde = s_dat
 
-    if isinstance(random_state, int):
+    if isinstance(random_state, (int,np.int_)):
         random_state = random_state + int(rep_values)
 
     xnew = skewnorm.rvs(
@@ -240,17 +241,18 @@ def _objective_two_sample(
         loc=mean_dat,
         scale=s_dat,
         a=skew_data,
-        random_state=random_state,
+        random_state=np.random.default_rng(random_state),
     )
     ynew = skewnorm.rvs(
         size=(m, len(mean_dat)),
         loc=mean_tilde,
         scale=s_tilde,
         a=skew_tilde,
-        random_state=random_state,
+        random_state=np.random.default_rng(random_state),
     )
 
-    statistic = stat_two_sample(xnew, ynew, h, np.array([[0]]), np.array([[1]]))
+    statistic = stat_two_sample(
+        xnew, ynew, h, np.array([[0]]), np.array([[1]]))
     cv = cv_twosample(
         num_iter,
         quantile,
@@ -368,7 +370,7 @@ def _objective_k_sample(
         skew_tilde = skew_data + dk
         s_tilde = s_dat
 
-    if isinstance(random_state, int):
+    if isinstance(random_state, (int,np.int_)):
         random_state = random_state + int(rep_values)
 
     nk = round(n / k)
@@ -522,7 +524,7 @@ def select_h(
     >>> print("Selected h is: ", h_selected)
     ... Selected h is:  1.2
     """
-    if not isinstance(random_state, (int, type(None))):
+    if not isinstance(random_state, (int, np.int_, type(None))):
         raise ValueError("Please specify a integer or None random_state")
 
     if isinstance(x, np.ndarray):
@@ -585,7 +587,8 @@ def select_h(
                 )
 
         mean_dat = np.mean(x, axis=0)
-        s_dat = np.diag(np.cov(x, rowvar=False).reshape(x.shape[1], x.shape[1]))
+        s_dat = np.diag(np.cov(x, rowvar=False).reshape(
+            x.shape[1], x.shape[1]))
         skew_data = skew(x)
         all_parameters = np.array(np.meshgrid(h_values, delta, rep_values)).T.reshape(
             -1, 3
@@ -611,22 +614,26 @@ def select_h(
                 )
                 for param in parameters
             )
-            results = pd.DataFrame(results, columns=["rep", "delta", "h", "score"])
+            results = pd.DataFrame(
+                results, columns=["rep", "delta", "h", "score"])
             results["score"] = 1 - results["score"]
             results_mean = (
-                results.groupby(["h", "delta"]).agg({"score": "mean"}).reset_index()
+                results.groupby(["h", "delta"]).agg(
+                    {"score": "mean"}).reset_index()
             )
             results_mean.columns = ["h", "delta", "power"]
             results_mean = results_mean.sort_values(by=["delta", "h"])
             all_results[delta_val] = results_mean
             min_h_power_gt_05 = results_mean[results_mean["power"] >= 0.5]
             if not min_h_power_gt_05.empty:
-                min_h = results_mean[results_mean["power"] >= 0.50].iloc[0]["h"]
+                min_h = results_mean[results_mean["power"]
[…]
