+++ tests/test_kernel_test.py
@@ -58,9 +58,9 @@ class TestKernelTest(unittest.TestCase):
         two_sample_test_subsampling = KernelTest(
             h=1.5, method="subsampling", b=0.5, random_state=42
         ).test(x, y)
-        self.assertTrue(two_sample_test_subsampling.un_h0_rejected_)
+        self.assertTrue(np.all(two_sample_test_subsampling.un_h0_rejected_))
         self.assertTrue(
-            isinstance(two_sample_test_subsampling.un_test_statistic_, (int, float))
+            isinstance(two_sample_test_subsampling.un_test_statistic_, np.ndarray)
         )
 
         with self.assertRaises(ValueError):
@@ -82,9 +82,9 @@ class TestKernelTest(unittest.TestCase):
             centering_type="param",
             random_state=42,
         ).test(x, y)
-        self.assertTrue(two_sample_test_h_selection.un_h0_rejected_)
+        self.assertTrue(np.all(two_sample_test_h_selection.un_h0_rejected_))
         self.assertTrue(
-            isinstance(two_sample_test_h_selection.un_test_statistic_, (int, float))
+            isinstance(two_sample_test_h_selection.un_test_statistic_, np.ndarray)
         )
 
     def test_ksample(self):
@@ -97,22 +97,22 @@ class TestKernelTest(unittest.TestCase):
         k_sample_test = KernelTest(
             h=1.5, method="subsampling", b=0.5, random_state=42
         ).test(x, y)
-        self.assertFalse(k_sample_test.un_h0_rejected_)
-        self.assertTrue(isinstance(k_sample_test.un_test_statistic_, (int, float)))
+        self.assertFalse(np.all(k_sample_test.un_h0_rejected_))
+        self.assertTrue(isinstance(k_sample_test.un_test_statistic_, np.ndarray))
 
         k_sample_test_without_h = KernelTest(
             method="subsampling", b=0.5, random_state=42, alternative="location"
         ).test(x, y)
-        self.assertFalse(k_sample_test_without_h.un_h0_rejected_)
+        self.assertFalse(np.all(k_sample_test_without_h.un_h0_rejected_))
         self.assertTrue(
-            isinstance(k_sample_test_without_h.un_test_statistic_, (int, float))
+            isinstance(k_sample_test_without_h.un_test_statistic_, np.ndarray)
         )
 
         k_sample_test_1d = KernelTest(
             h=1.5, method="subsampling", b=0.5, random_state=42
         ).test(x_1d, y_dataframe)
-        self.assertFalse(k_sample_test_1d.un_h0_rejected_)
-        self.assertTrue(isinstance(k_sample_test_1d.un_test_statistic_, (int, float)))
+        self.assertFalse(np.all(k_sample_test_1d.un_h0_rejected_))
+        self.assertTrue(isinstance(k_sample_test_1d.un_test_statistic_, np.ndarray))
 
         with self.assertRaises(ValueError):
             KernelTest(h=1.5, method="subsampling", b=0.5, random_state=42).test(
