### 2024-05-06 666a302
Updated basic_usage and black formatting
