### 2024-05-13 6255b45
Bumped version from dev to v1.1.0
Updated variance computations, included additional usage examples
