+++ doc/source/changelog/index.rst
@@ -6,5 +6,6 @@ Changelogs and release notes for all QuadratiK releases are linked in this page.
 .. toctree::
    :maxdepth: 1
 
+   v1.1.0.rst
    v1.0.0.rst
-   v1.0.1.rst
+   
